refactor: move GraphCSR to e3j/data
